% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_QSAR.R
\name{extract_features_QSAR}
\alias{extract_features_QSAR}
\title{Extract QSAR Features from Peptide Sequences}
\usage{
extract_features_QSAR(
  df = NULL,
  n = NULL,
  sequence_col = "Sequence",
  docking_col = NULL,
  pH = 7.4,
  normalize = FALSE
)
}
\arguments{
\item{df}{A data frame or a vector of peptide sequences. If 'df' is provided, 'n' will be ignored.}

\item{n}{An integer representing the length of the peptide library to be generated. If 'df' is provided, 'n' will be ignored.}

\item{sequence_col}{A string representing the name of the column containing the peptide sequences.}

\item{docking_col}{A string representing the name of the column containing the docking information.}

\item{pH}{The pH used for calculating charge (default is 7.4).}

\item{normalize}{A boolean indicating if the data should be normalized (default is FALSE).}
}
\value{
A dataframe with the calculated peptide properties.
}
\description{
This function extracts various Quantitative Structure-Activity Relationship (QSAR) features
from peptide sequences. The extraction is based on a variety of amino acid properties
and functions from the "Peptides" package (https://github.com/dosorio/Peptides/).
}
\examples{
extract_features_QSAR(df = c('ACA', 'EDE'))
}
