% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_seq.R
\name{increment}
\alias{increment}
\title{Increment Peptide Sequences}
\usage{
increment(peptides, num_added = 1)
}
\arguments{
\item{peptides}{A character vector of peptide sequences.}

\item{num_added}{The number of amino acids to be added to each position of the peptide.}
}
\value{
A character vector of new peptide sequences.
}
\description{
This function generates new peptide sequences by adding each of the 20 amino acids
to each position of the input peptide or peptides.
}
\examples{
increment(c("AC", "DE"))
increment("ACDE", num_added = 2)
}
